requirements.txt
requests~=2.25.0
psycopg2~=2.8.6
discord~=1.0.1
python-dateutil~=2.8.1
beautifulsoup4~=4.9.3
Pillow
smashggpy~=0.0.29